requirements.txt
bitcoin
accrocchio
siringa
websocket-client
